chore: #83 UserInfo 디렉토리 변경
